Update method to pass HTML in Puppeteer provider (#25)
The previous method used to pass HTML to the browser in the example Puppeteer provider truncates output if certain characters are present in the HTML, including `#` which is present in HTML entities. `page.setContent` has been able to accept a `waitUntil` option for the last couple of years so this is now an available approach.